+++ CHANGELOG.md
@@ -53,6 +53,7 @@
 - directory detection of completion wrapper
 - return type resolution of anonymous function that last statement is Nothing type
 - file name completion in prefix assignment
+- pass invalid ``SCRIPT_DIR`` in module name completion
 
 ## [0.25.0] - 2022-03-31